rename qc dir to src
